+++ Sources/BezierSwift/MasterComponent/BezierToast/BezierToastContainerView.swift
@@ -5,22 +5,7 @@
 //  Created by woody on 2023/03/31.
 //
 
-import Foundation
 import SwiftUI
-
-import Combine
-
-struct BezierToastItem: Identifiable {
-  let param: BezierToastParam
-  let binding: Binding<BezierToastParam?>?
-  let uuid = UUID()
-  var id: UUID { self.uuid }
-  
-  init(param: BezierToastParam, binding: Binding<BezierToastParam?>? = nil) {
-    self.param = param
-    self.binding = binding
-  }
-}
 
 private enum Metric {
   static let topPadding = CGFloat(5)
